[:hammer: refactor] mueve las funciones de la ruta principal
