+++ _funcs/getSO.py
